# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PAOSITRA - API de gestion",
    "description": "API versionnée des lots Trésorerie et Gestion des opérations. Les choix techniques documentés ne constituent pas des exigences contractuelles du DAO.",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/treasury/budget/versions": {
      "get": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Versions d'un exercice budgetaire",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Versions budgetaires"
          }
        }
      },
      "post": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Creation ou copie d'une version budgetaire",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Version brouillon creee"
          }
        }
      }
    },
    "/api/v1/treasury/budget/versions/{id}/activate": {
      "post": {
        "tags": [
          "Tresorerie"
        ],
        "summary": "Activation exclusive d'une version budgetaire",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Version activee"
          }
        }
      }
    },
...
  }
}